Add base Card scene
Includes some scaling of font size for the card labels, so that display
of bigger numbers for attack/defense/stamina looks better.

However, includes a minimum font size so that text don't entirely
disappear. Currently, the "Chaos Slash" card's description still runs
off the card, but that seems like we just need to shorten the
description and have a better UI for detailed card effects.